requirements.txt
pandas
numpy
scikit-learn
matplotlib
python-dotenv
pymongo
certifi
pymongo[srv]==3.12
##-e .